Invert reset signal when exit.

+++ src/modem-manager.c
@@ -94,13 +94,6 @@ static inline gboolean pcat_modem_manager_modem_power_init(
         return FALSE;
     }
 
-    if(main_config_data->hw_gpio_modem_rf_kill_chip==NULL)
-    {
-        g_warning("Modem RF kill GPIO chip not configured!");
-
-        return FALSE;
-    }
-
     if(main_config_data->hw_gpio_modem_reset_chip==NULL)
     {
         g_warning("Modem reset GPIO chip not configured!");
@@ -120,15 +113,15 @@ static inline gboolean pcat_modem_manager_modem_power_init(
         }
     }
 
-    if(mm_data->gpio_modem_rf_kill_chip==NULL)
+    if(mm_data->gpio_modem_rf_kill_chip==NULL &&
+        main_config_data->hw_gpio_modem_rf_kill_chip!=NULL &&
+        *(main_config_data->hw_gpio_modem_rf_kill_chip)!='\0')
     {
         mm_data->gpio_modem_rf_kill_chip = gpiod_chip_open_by_name(
             main_config_data->hw_gpio_modem_rf_kill_chip);
         if(mm_data->gpio_modem_rf_kill_chip==NULL)
         {
-            g_warning("Failed to open Modem power GPIO chip!");
-
-            return FALSE;
+            g_warning("Failed to open Modem RF Kill GPIO chip!");
         }
     }
 
@@ -172,7 +165,8 @@ static inline gboolean pcat_modem_manager_modem_power_init(
             main_config_data->hw_gpio_modem_power_active_low ? 1 : 0);
     }
 
-    if(mm_data->gpio_modem_rf_kill_line==NULL)
+    if(mm_data->gpio_modem_rf_kill_line==NULL &&
+        mm_data->gpio_modem_rf_kill_chip!=NULL)
     {
         mm_data->gpio_modem_rf_kill_line = gpiod_chip_get_line(
             mm_data->gpio_modem_rf_kill_chip,
@@ -180,24 +174,26 @@ static inline gboolean pcat_modem_manager_modem_power_init(
         if(mm_data->gpio_modem_rf_kill_line==NULL)
         {
             g_warning("Failed to open Modem RF kill GPIO line!");
+        }
+    }
 
-            return FALSE;
-        }
-    }
-    if(!gpiod_line_is_requested(mm_data->gpio_modem_rf_kill_line))
+    if(mm_data->gpio_modem_rf_kill_line!=NULL)
     {
-        ret = gpiod_line_request_output(mm_data->gpio_modem_rf_kill_line,
-            "gpio-modem-rf-kill",
-            main_config_data->hw_gpio_modem_rf_kill_active_low ? 0 : 1);
-        if(ret!=0)
+        if(!gpiod_line_is_requested(mm_data->gpio_modem_rf_kill_line))
         {
-            g_warning("Failed to request output on Modem RF kill GPIO!");
+            ret = gpiod_line_request_output(mm_data->gpio_modem_rf_kill_line,
+                "gpio-modem-rf-kill",
+                main_config_data->hw_gpio_modem_rf_kill_active_low ? 0 : 1);
+            if(ret!=0)
+            {
+                g_warning("Failed to request output on Modem RF kill GPIO!");
+            }
         }
-    }
-    else
-    {
-        gpiod_line_set_value(mm_data->gpio_modem_rf_kill_line,
-            main_config_data->hw_gpio_modem_rf_kill_active_low ? 0 : 1);
+        else
+        {
+            gpiod_line_set_value(mm_data->gpio_modem_rf_kill_line,
+                main_config_data->hw_gpio_modem_rf_kill_active_low ? 0 : 1);
+        }
     }
 
     if(mm_data->gpio_modem_reset_line==NULL)
@@ -239,8 +235,11 @@ static inline gboolean pcat_modem_manager_modem_power_init(
 
     gpiod_line_set_value(mm_data->gpio_modem_power_line,
         main_config_data->hw_gpio_modem_power_active_low ? 0 : 1);
-    gpiod_line_set_value(mm_data->gpio_modem_rf_kill_line,
-        main_config_data->hw_gpio_modem_rf_kill_active_low ? 1 : 0);
+    if(mm_data->gpio_modem_rf_kill_line!=NULL)
+    {
+        gpiod_line_set_value(mm_data->gpio_modem_rf_kill_line,
+            main_config_data->hw_gpio_modem_rf_kill_active_low ? 1 : 0);
+    }
     gpiod_line_set_value(mm_data->gpio_modem_reset_line,
         main_config_data->hw_gpio_modem_reset_active_low ? 1 : 0);
 
@@ -759,7 +758,7 @@ static gpointer pcat_modem_manager_modem_work_thread_func(
     if(mm_data->gpio_modem_reset_line!=NULL)
     {
         gpiod_line_set_value(mm_data->gpio_modem_reset_line,
-            main_config_data->hw_gpio_modem_reset_active_low ? 0 : 1);
+            main_config_data->hw_gpio_modem_reset_active_low ? 1 : 0);
 
         gpiod_line_release(mm_data->gpio_modem_reset_line);
         mm_data->gpio_modem_reset_line = NULL;
